feat: seperate grpc logic to a new location
